refactor: 수익률 계산의 책임을 StatisticCalculator 로 위임 및 리턴 타입을 변경
- 데이터만을 리턴하도록 dto mapper 가 만든 dto 로 변경
- `통계`와 관련된 계산 책임을 지키도록 메서드 위치를 이동
