+++ src/main/java/lotto/dto/StatisticDto.java
@@ -1,0 +1,7 @@
+package lotto.dto;
+
+import java.util.EnumMap;
+import lotto.domain.Rank;
+
+public record StatisticDto(EnumMap<Rank, Integer> statisticSummary, float profitRate) {
+}
